backport to Minecraft version 1.20.1 also updated dependencies

+++ src/main/kotlin/de/jagenka/PerformanceMetrics.kt
@@ -1,3 +1,3 @@
 package de.jagenka
 
-data class PerformanceMetrics(val mspt: Double, val tps: Float)
+data class PerformanceMetrics(val mspt: Double, val tps: Double)
